Fix changes to month dropdown
When using the month dropdown option changes don’t take effect because
moment(opt[type])[name](value) is passed a string. The momentJS docs
state that this works when name===‘month’ but testing reveals that this
isn’t the case. Always passing a Number instead of a String always
works.